Renaming UI files for consistency - adjustment
