+++ ui/main/src/main.ts
@@ -63,7 +63,7 @@ import {TagInputModule} from 'ngx-chips';
 import {OAuthModule} from 'angular-oauth2-oidc';
 import {NgbModule, NgbModalModule} from '@ng-bootstrap/ng-bootstrap';
 import {TranslateModule} from '@ngx-translate/core';
-import {AppRoutingModule} from './app/services/navigation/router/AngularApplicationRouter.module';
+import {AppRoutingModule} from './app/services/navigation/router/AngularApplicationRouterModule';
 import {AppComponent} from './app/AppComponent';
 import {NgxDaterangepickerMd} from 'ngx-daterangepicker-material';
 import {AngularHandlebarsTemplateServer} from '@ofServices/handlebars/server/AngularHandlebarsTemplateServer';
